pyproject.toml
[project]
name = "crag-pipeline"
version = "0.1.0"
description = "Corrective RAG Pipeline with LangGraph, Qdrant, and LangSmith"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-qdrant>=0.1.0",
    "qdrant-client>=1.12.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "fastembed>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-magic-bin>=0.4.14",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "watchdog>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "black>=24.10.0",
    "isort>=5.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pre-commit>=4.0.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
    "black>=24.10.0",
    "isort>=5.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pre-commit>=4.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C", "T"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
